Raz's Homework 02

+++ OneOfEachStats.java
@@ -50,7 +50,7 @@ public class OneOfEachStats {
 				numOf4Children++;
 			}
 		}
-		System.out.println("Average: " + (int)(countTotal/T) + " children to get at least one of each gender.");
+		System.out.println("Average: " + (double)(countTotal/T) + " children to get at least one of each gender.");
 		System.out.println("Number of families with 2 children: " + numOf2Children);
 		System.out.println("Number of families with 3 children: " + numOf3Children);
 		System.out.println("Number of families with 4 or more children: " + numOf4Children);
